Further decouple `Calendar` from GraphQL result types
Previously the `_repoDay()` and `cleanRepository()` methods took
parameters of type `gql.Repository`. This adds a new type,
`RepositorySource` that can be either a string (for convenience when
writing tests) or an object with the fields that both `model.Repository`
and `gql.Repository` have.

This also allows for the removal of the second `repoDay()` method,
making things less confusing.