Expand spacing
Most spacing (`line-height`, `padding`, `margin`, etc.) was specified in
`rem`, which is the font size of the root element. That made copy feel
cramped, since it was using the font size for the leading.

This switches from `rem` to `rlh`, the leading on the root element. Done
entirely mechanically with:

    /(\d+\s*)rem\b/ -> \1rlh